requirements.txt
langchain-community==0.0.37
langchain-core==0.1.51
langchain-openai==0.0.6
langchain-pinecone==0.1.0
python-dotenv==1.0.1

PyYAML==6.0.1
fastapi==0.111.0
uvicorn==0.29.0
datasets==2.19.1
pydantic==2.7.1

pytest==8.2.0